style: cleanup
